customModeltf/requirements.txt
fastapi
uvicorn
tensorflow
Pillow
python-multipart